Apply review feedback: i18n improvements
