+++ pages/blog/blog-app.js
@@ -10,7 +10,7 @@ import React from 'https://esm.sh/react@19.0.0';
 import { createRoot } from 'https://esm.sh/react-dom@19.0.0/client';
 import htm from 'https://esm.sh/htm@3.1.1';
 import { createLogger } from '/content/core/logger.js';
-import { i18n } from '/content/core/i18n.js';
+import { useTranslation } from '/content/core/hooks/useTranslation.js';
 import { updateLoader, hideLoader } from '/content/core/global-loader.js';
 import { marked } from 'https://cdn.jsdelivr.net/npm/marked@11.1.1/lib/marked.esm.js';
 import DOMPurify from 'https://cdn.jsdelivr.net/npm/dompurify@3.0.8/dist/purify.es.mjs';
@@ -20,19 +20,6 @@ marked.setOptions({ mangle: false, headerIds: false });
 
 const log = createLogger('BlogApp');
 const html = htm.bind(React.createElement);
-
-// Translation Hook
-const useTranslation = () => {
-  const [lang, setLang] = React.useState(i18n.currentLang);
-
-  React.useEffect(() => {
-    const onLangChange = (e) => setLang(e.detail.lang);
-    i18n.addEventListener('language-changed', onLangChange);
-    return () => i18n.removeEventListener('language-changed', onLangChange);
-  }, []);
-
-  return { t: (key, params) => i18n.t(key, params), lang };
-};
 
 // --- Utilities ---
 const estimateReadTime = (text = '') =>
